\name{cij.play.1}
\alias{cij.play.1}
\docType{data}
\title{
Small cij matrix to test dccm.mean
}
\description{
A 4x4 matrix useful for tests. }
\usage{data(cij.play.1)}
\format{
  The format is:
 num [1:4, 1:4] 1 0 0 0.5 1 1 0.5 0.5 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "V1" "V2" "V3" "V4"
}
\examples{
data(cij.play.1)
## maybe str(cij.play.1) ; plot(cij.play.1) ...
}
\keyword{datasets}
