\name{cij.play.2}
\alias{cij.play.2}
\docType{data}
\title{
Small cij matrix to test dccm.mean.
}
\description{
A 4x4 matrix useful for tests.
}
\usage{data(cij.play.2)}
\format{
  The format is:
 num [1:4, 1:4] 1 0 0 0 1 1 0 0 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "V1" "V2" "V3" "V4"
}
\examples{
data(cij.play.2)
## maybe str(cij.play.2) ; plot(cij.play.2) ...
}
\keyword{datasets}
