\name{d.autonomy}

\docType{data}

\alias{d.autonomy}

\title{Emergence and endurance of autonomy of biodiversity institutions in Costa Rica}

\description{
This dataset is from Basurto (2013), who analyzes the causes of the emergence and endurance of autonomy among local institutions for biodiversity conservation in Costa Rica between 1986 and 2006. 
}

\usage{d.autonomy}

\format{
The data frame contains 30 rows (cases), which are divided in two halves: rows 1 to 14 comprise data on the emergence of local autonomy between 1986 and 1998, rows 15 to 30 comprise data on the endurance of local autonomy between 1998 and 2006. The data has the following 9 columns featuring fuzzy-set factors:

\tabular{rlll}{
[ , 1] \tab \bold{AU}  \tab local autonomy  (ultimate outcome)                                        \cr
[ , 2] \tab \bold{EM}  \tab local communal involvement through direct employment                       \cr
[ , 3] \tab \bold{SP}  \tab  local direct spending                                          \cr                          
[ , 4] \tab \bold{CO}  \tab co-management with local or regional stakeholders                          \cr
[ , 5] \tab \bold{CI}  \tab degree of influence of national civil service policies         \cr
[ , 6] \tab \bold{PO}  \tab  national participation in policy-making                        \cr
[ , 7] \tab \bold{RE}  \tab  research-oriented partnerships                                 \cr   
[ , 8] \tab \bold{CN}  \tab conservation-oriented partnerships                             \cr
[ , 9] \tab \bold{DE}  \tab  direct support by development organizations              
}
}


\section{Contributors}{
Thiem, Alrik: collection, documentation
}

\source{
Basurto, Xavier. 2013. \dQuote{Linking Multi-Level Governance to Local Common-Pool 
Resource Theory using Fuzzy-Set Qualitative Comparative Analysis: Insights from 
Twenty Years of Biodiversity Conservation in Costa Rica.} \emph{Global Environmental Change} \strong{23} (3):573-87. 
}

\keyword{datasets}
