\name{d.educate}

\docType{data}

\alias{d.educate}

\title{Artificial data on education levels and left-party strength}

\description{
This artificial dataset of macro-sociological factors on high levels of education is from Baumgartner (2009).
}

\usage{d.educate}

\format{
The data frame contains 8 rows (cases) and the following 5 columns featuring Boolean factors taking values 1 and 0 only:

\tabular{rll}{
[ , 1] \tab \bold{U} \tab existence of strong unions      \cr
[ , 2] \tab \bold{D} \tab high level of disparity         \cr
[ , 3] \tab \bold{L} \tab existence of strong left parties\cr
[ , 4] \tab \bold{G} \tab high gross national product     \cr
[ , 5] \tab \bold{E} \tab high level of education
}
}

\source{
Baumgartner, Michael. 2009. \dQuote{Inferring Causal Complexity.}
\emph{Sociological Methods & Research} 38(1):71-101.
}

\keyword{datasets}