\name{d.irrigate}

\docType{data}

\alias{d.irrigate}

\title{Data on the impact of development interventions on water adequacy in Nepal}

\description{
This dataset is from Lam and Ostrom (2010), who analyze the effects of an irrigation
experiment in Nepal.
}

\usage{d.irrigate}

\format{
The dataset contains 15 rows (cases) and the following 6 columns featuring Boolean factors taking values 1 and 0 only:

\tabular{rll}{
[ , 1] \tab \bold{A} \tab continual assistance on infrastructure improvement                         \cr
[ , 2] \tab \bold{R} \tab existence of a set of formal rules for irrigation operation and maintenance\cr
[ , 3] \tab \bold{F} \tab existence of provisions of fines                                           \cr
[ , 4] \tab \bold{L} \tab existence of consistent leadership                                         \cr
[ , 5] \tab \bold{C} \tab existence of collective action among farmers for system maintenance        \cr
[ , 6] \tab \bold{W} \tab persistent improvement in water adequacy at the tail end in winter
}
}

\source{
Lam, Wai Fung, and Elinor Ostrom. 2010.
\dQuote{Analyzing the Dynamic Complexity of Development Interventions: Lessons
from an Irrigation Experiment in Nepal.}
\emph{Policy Sciences} 43 (2):1-25.
}

\keyword{datasets}