\name{d.minaret}

\docType{data}

\alias{d.minaret}

\title{Data on the voting outcome of the 2009 Swiss Minaret Initiative}

\description{
This dataset is from Baumgartner and Epple (2014), who analyze the determinants
of the outcome of the vote on the 2009 Swiss Minaret Initative.
}

\usage{d.minaret}

\format{
The data frame contains 26 rows (cases) and the following 6 columns featuring raw data:

\tabular{rll}{
[ , 1] \tab \bold{A} \tab rate of old xenophobia                                    \cr
[ , 2] \tab \bold{L} \tab left party strength                                       \cr
[ , 3] \tab \bold{S} \tab share of native speakers of Serbian, Croatian, or Albanian\cr
[ , 4] \tab \bold{T} \tab strength of traditional economic sector                   \cr
[ , 5] \tab \bold{X} \tab rate of new xenophobia                                    \cr
[ , 6] \tab \bold{M} \tab acceptance of Minaret Initiative
}
}

\section{Contributors}{
Ruedi Epple: collection, documentation
}

\source{
Baumgartner, Michael, and Ruedi Epple. 2014. \dQuote{A Coincidence Analysis of a
Causal Chain: The Swiss Minaret Vote.}
\emph{Sociological Methods & Research} 43 (2):280-312.
}

\keyword{datasets}