\name{d.pacts}

\docType{data}

\alias{d.pacts}

\title{Data on the emergence of labor agreements in new democracies between 1994 and 2004}

\description{
This dataset is from Aleman (2009), who analyzes the causes of the emergence of tripartite labor agreements among unions, employers, and government representatives in new democracies in Europe, Latin America, Africa, and Asia between 1994 and 2004.
}

\usage{d.pacts}

\format{
The data frame contains 78 rows (cases) and the following 5 columns listing membership scores in 5 fuzzy sets:

\tabular{rll}{
[ , 1] \tab \bold{PACT} \tab development of tripartite cooperation (ultimate outcome)\cr
[ , 2] \tab \bold{W} \tab regulation of the wage setting process\cr
[ , 3] \tab \bold{E} \tab regulation of the employment process                                          \cr
[ , 4] \tab \bold{L} \tab presence of a left government                                         \cr
[ , 5] \tab \bold{P} \tab presence of an encompassing labor organization (labor power)        
}
}

\section{Contributors}{
Thiem, Alrik: collection, documentation
}

\source{
Aleman, Jose. 2009.
\dQuote{The Politics of Tripartite Cooperation in New Democracies: A Multi-level Analysis.}
\emph{International Political Science Review} 30 (2):141-162.
}

\keyword{datasets}