\name{d.performance}

\docType{data}

\alias{d.performance}

\title{Data on combinations of industry, corporate, and business-unit effects}

\description{
This dataset is from Greckhammer et al. (2008), who analyze the causal conditions for superior (above average) business-unit performance of corporations in the manufacturing sector during the years 1995 to 1998.
}

\usage{d.performance}

\format{
The data frame contains 214 rows featuring configurations, one column reporting the frequencies of each configuration, and 8 columns listing the following Boolean factors:

\tabular{rll}{
[ , 1] \tab \bold{MU} \tab above average industry munificence     \cr
[ , 2] \tab \bold{DY} \tab high industry dynamism \cr
[ , 3] \tab \bold{CO} \tab high industry competitiveness   \cr
[ , 4] \tab \bold{DIV} \tab high corporate diversification      \cr
[ , 5] \tab \bold{CRA} \tab above median corporate resource availability        \cr
[ , 6] \tab \bold{CI} \tab  above median corporate capital intensity \cr 
[ , 7] \tab \bold{BUS} \tab  large business-unit size \cr
[ , 8] \tab \bold{SP} \tab  above average business-unit performance (in the manufacturing sector)\cr
}
}

\source{
Greckhamer, Thomas, Vilmos F. Misangyi, Heather Elms, and Rodney Lacey. 2008.
\dQuote{Using Qualitative Comparative Analysis in Strategic Management Research: An Examination of Combinations of Industry, Corporate, and Business-Unit Effects.}
\emph{Organizational Research Methods} 11 (4):695-726.
}

\keyword{datasets}