\name{d.volatile}

\docType{data}

\alias{d.volatile}

\title{Data on the volatility of grassroots associations in Norway between 1980 and 2000}

\description{
This dataset is from Wollebaek (2010), who analyzes the causes of disbandings of grassroots associations in Norway.
}

\usage{d.volatile}

\format{
The data frame contains 22 rows (cases) and the following 9 columns featuring Boolean factors taking values 1 and 0 only:

\tabular{rll}{
[ , 1] \tab \bold{PG} \tab high population growth                         \cr
[ , 2] \tab \bold{RB} \tab high rurbanization (i.e. people moving to previously sparsely populated areas that are\cr\tab\tab not adjacent to a larger city)\cr
[ , 3] \tab \bold{EL} \tab high increase in education levels                                           \cr
[ , 4] \tab \bold{SE} \tab high degree of secularization                                    \cr
[ , 5] \tab \bold{CS} \tab existence of Christian strongholds        \cr
[ , 6] \tab \bold{OD} \tab high organizational density \cr
[ , 7] \tab \bold{PC} \tab existence of polycephality (i.e. municipalities with multiple centers)\cr
[ , 8] \tab \bold{UP} \tab urban proximity        \cr
[ , 9] \tab \bold{VO2} \tab very high volatility of grassroots associations        
}
}

\source{
Wollebaek, Dag. 2010.
\dQuote{Volatility and Growth in Populations of Rural Associations.}
\emph{Rural Sociology} 75:144-166.
}

\keyword{datasets}