\name{d.women}

\docType{data}

\alias{d.women}

\title{Data on high percentage of women's representation in parliaments of western countries}

\description{
This dataset is from Krook (2010), who analyzes the causal conditions for high women's representation in western-democratic parliaments.
}

\usage{d.women}

\format{
The data frame contains 22 rows (cases) and the following 6 columns featuring Boolean factors taking values 1 and 0 only:
\tabular{rll}{
[ , 1] \tab \bold{ES} \tab existence of a PR electoral system     \cr
[ , 2] \tab \bold{QU} \tab existence of quotas for women \cr
[ , 3] \tab \bold{WS} \tab existence of social-democratic welfare system   \cr
[ , 4] \tab \bold{WM} \tab existence of autonomous women's movement      \cr
[ , 5] \tab \bold{LP} \tab strong left parties        \cr
[ , 6] \tab \bold{WNP} \tab high women's representation in parliament 
}
}

\source{
Krook, Mona Lena. 2010.
\dQuote{Women's Representation in Parliament: A Qualitative Comparative Analysis.}
\emph{Political Studies} 58 (5):886-908.
}

\keyword{datasets}