\name{trj.1}
\alias{trj.1}
\alias{trj.2}
\alias{trj.3}
\docType{data}
\title{
Example Trajctory Files
}
\description{
A numeric matrix (mXn), where rows are the trajectory frames and columns the x, y and z atomic coordinates.
}
\usage{data(trj.1)}
\format{
A numeric matrix mXn matrix
}
\examples{
#data(trj.1)
#Read PDB and do all sorts of analysis!
}
\keyword{datasets}
