\name{CNA-package}
\alias{CNA}
\alias{CNA}
\docType{package}
\title{
Contact Network Assortativeness
}
\description{
This package calculates assortativeness and bridging behaviours of a network.

}
\details{
\tabular{ll}{
Package: \tab CNA\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-10-15\cr
License: \tab \cr
}
 \code{
 \link{assort.boots}\cr
 \link{assort.Farrington}\cr
 \link{assort.Gupta}\cr
  \link{assort.jackknife}\cr
 \link{assort.Newman}\cr
 \link{boot.assort}\cr
 \link{bridge}\cr
 \link{cut.cna}\cr
 \link{find.component}\cr
 \link{sampling}\cr
 \link{sampling.method}
 }
}
\author{
Fei MENG \email{fei.d.meng@gmail.com}
}

\references{
Gupta S, Anderson RM, May RM (1989) Networks of sexual contacts: implications for the pattern of spread of HIV. AIDS 3: 807-817.,

Newman ME (2003) Mixing patterns in networks. Phys Rev E Stat Nonlin Soft Matter Phys 67: 026126.,

Farrington CP, Whitaker HJ, Wallinga J, Manfredi P (2009) Measures of disassortativeness and their application to directly transmitted infections. Biom J 51: 387-407.
}

\keyword{ network assortativity }
\seealso{
\pkg{network}
\pkg{sna}
}
\examples{ 

}
