\name{d.educate}

\docType{data}

\alias{d.educate}

\title{Artifical data on education levels and left-party strength}

\description{
This artifical dataset of macrosociological factors on high levels of education is from Baumgartner (2009).
}

\usage{data(d.educate)}

\format{
This dataset contains 8 rows (cases) and the following 5 columns (Boolean factors):

\tabular{rll}{
[ , 1] \tab \bold{U} \tab existence of strong unions      \cr
[ , 2] \tab \bold{D} \tab high level of disparity         \cr
[ , 3] \tab \bold{L} \tab existence of strong left parties\cr
[ , 4] \tab \bold{G} \tab high gross national product     \cr
[ , 5] \tab \bold{E} \tab high levels of education
}
}

\source{
Baumgartner, Michael. 2009. \dQuote{Inferring Causal Complexity.}
\emph{Sociological Methods & Research} 38(1):71-101.
}

\keyword{datasets}