\name{d.minaret}

\docType{data}

\alias{d.minaret}

\title{Data on the voting outcome of the 2009 Swiss Minaret Initiative}

\description{
This dataset is from Baumgartner and Epple (2014), who analyze the determinants
of the outcome of the vote on the 2009 Swiss Minaret Initative.
}

\usage{data(d.minaret)}

\format{
This dataset contains 26 rows (cases) and the following 6 columns (variables):

\tabular{rll}{
[ , 1] \tab \bold{a} \tab rate of old xenophobia                                    \cr
[ , 2] \tab \bold{l} \tab left party strength                                       \cr
[ , 3] \tab \bold{s} \tab share of native speakers of Serbian, Croatian, or Albanian\cr
[ , 4] \tab \bold{t} \tab strength of traditional economic sector                   \cr
[ , 5] \tab \bold{x} \tab rate of new xenophobia                                    \cr
[ , 6] \tab \bold{m} \tab acceptance of Minaret Initiative
}
}

\source{
Baumgartner, Michael, and Ruedi Epple. 2014. \dQuote{A Coincidence Analysis of a
Causal Chain: The Swiss Minaret Vote.}
\emph{Sociological Methods & Research} 43 (2):280-312.
}

\keyword{datasets}